requirements_pip.txt
absl-py==2.0.0
accelerate==0.19.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip==1.0
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
datasets==2.16.1
decorator==5.1.1
diffusers==0.17.0
dill==0.3.6
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.108.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
future==0.18.3
google-auth==2.25.2
google-auth-oauthlib==1.0.0
gradio==3.36.1
gradio-client==0.2.7
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.6
imageio==2.31.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
install==1.3.5
ipython==8.12.3
ipywidgets==8.1.1
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kornia==0.7.1
lazy-loader==0.3
lightning-utilities==0.10.0
linkify-it-py==2.0.2
markdown==3.5.1
markdown-it-py==2.2.0
markupsafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.1.1
open-clip-torch==2.23.0
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.1.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic-core==2.14.6
pydeprecate==0.3.1
pydub==0.25.1
pygments==2.17.2
pyparsing==3.1.1
pyre-extensions==0.0.23
python-box==6.0.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.4.2
pytz==2023.3.post1
pywavelets==1.4.1
pyyaml==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rpds-py==0.15.2
rsa==4.9
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.1
semantic-version==2.10.0
sentencepiece==0.1.99
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.32.0.post1
sympy==1.12
taming-transformers==0.0.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tokenizers==0.13.3
toolz==0.12.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchmetrics==0.6.0
torchvision==0.14.1+cu117
tqdm==4.66.1
traitlets==5.14.1
transformers==4.28.1
triton==2.1.0
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.1.0
uvicorn==0.25.0
wcwidth==0.2.12
websockets==11.0.3
werkzeug==3.0.1
widgetsnbextension==4.0.9
zipp==3.18.0
